imx8m: soc.c: Add vendor_board_fix_fdt
NXP added the board_fix_fdt() function to the soc.c file, which is not ideal.
If a specific board needs its own board_fix_fdt implementation, it would
override NXP's function, potentially causing unexpected issues.

To address this, the __weak vendor_board_fix_fdt() function is introduced.
It is always called at the end of board_fix_fdt and can be redefined in
vendor-specific board files as needed.

Signed-off-by: Alifer Moraes <alifer.m@variscite.com>